datastreams: rename fixtures
